Rename util methods and create TimeUtil
