+++ src/main/java/net/dv8tion/jda/internal/entities/GuildImpl.java
@@ -703,7 +703,7 @@ public class GuildImpl implements Guild
                 if (ChronoUnit.MINUTES.between(getSelfMember().getTimeJoined(), OffsetDateTime.now()) < 10)
                     break;
             case MEDIUM:
-                if (ChronoUnit.MINUTES.between(MiscUtil.getCreationTime(getJDA().getSelfUser()), OffsetDateTime.now()) < 5)
+                if (ChronoUnit.MINUTES.between(getJDA().getSelfUser().getTimeCreated(), OffsetDateTime.now()) < 5)
                     break;
             case LOW:
                 if (!getJDA().getSelfUser().isVerified())
